test_requirements.txt
colour-runner==0.0.4
django-discover-runner==1.0
